best_practices/contrib/depression_detection/inference/requirements.txt
tf2onnx==1.12.1
tensorflow==2.6.0
Scikit-learn==0.24.0
numpy==1.19.5
pandas==1.1.5
onnxruntime==1.10.0
keras==2.6.0
django==3.1.2
onnxmltools==1.11.1
packaging==21.3